hackathon_chat_test/server/requirements.txt
Flask==2.2.3
Flask-Cors==3.0.10
Werkzeug==2.2.3
python-dotenv==1.0.0
langchain==0.1.0
openai==1.7.1
#faiss
#pdfreader
#psycopg2